TestInitializationListener skeleton
Summary:
This is skeleton of TestInitializationListener which is responsible for
fulfilling test environment for whole test suite and individual tests.
This is not working yet - but shows overal structure of test suite, test
case execution flow.

Next step would be to eliminate all the todos present in code, and
probably some more :)

Test Plan: Build

Reviewers: ksobczak, anton

Subscribers: sogorkis

Differential Revision: https://sdvl3bdch001.td.teradata.com/D6