build: introduce third-party-deps project to store provided dependencies
Summary:
The third-party-deps project will hold dependencies that cannot/should not be
added to the test framework core fat jar. Those dependencies are
marked as provided and are ignored when building the fat jar.

For Hadoop core and Presto JDBC we just copy those jars over while the Hive JDBC
driver we have to package in a fat jar. The Hive JDBC driver fat jar
will not be placed on the classpath and instead will be loaded at runtime
in a separate classloader in order to not cause a conflict with a
newer version of Guava. Hive JDBC depends on Hadoop at runtime which is why
the Hadoop core jar is there.

Task: SWARM-216
Review Url: @@review_url@@

Test Plan: Ran all example tests

Differential Revision: https://phabricator.td.teradata.com/D111