Add support for allOf in tests execution
Summary:
This patch adds support for execution of tests which use allOf
clause in requirements definition.

Implementation is quite hacky (partly due to TestNG weirdnesses).
We add new intercepting listener (RequirementsExtender) whose purpose is
to calculate requirement sets list of test methods. For each input test
method one or more result test methods is created. Result test methods
are not standard TestNGMethod but RequirementsAwareTestNGMethod which
contain extra field with set of requirements needed by test method
instance.

For some reason interceptor is called after onStart method of
TestInitializationListener - so TestInitializationListener have to
calculate list of requirements of all tests manually.

We also add number of tests after interception as attribute in
testContext as for some reason list of all tests to be run in test
contexts is not updated by testng, and we cannot use that list in
ProgressLoggingListener to compute number of all tests to be run.

Test Plan: Build, example tests

Reviewers: ksobczak, sogorkis

Differential Revision: https://phabricator.td.teradata.com/D352