Support for convetion test cases that are packaged in JAR file.
Summary:
Now we do not use File in our convention code, since it does not
support >Java7 FileSystems. We use Path instead.
Added a method, processPathFromUri that creates a Path object from
URI and (when required) opens a FileSystem for that URI for the duration
of processing. This way we support convention tests cases in Jar files.

Task: SWARM-246

Test Plan: manual, unit

Reviewers: wbiela, sogorkis

Differential Revision: https://phabricator.td.teradata.com/D377