core: support multiple jdbc table managers
Now user can specify multiple table managers on the same database. This
is particular useful for JdbcTableManger. See tempto-configuration.yaml
in tempto-examples.

ImmutableTablesState had to change, before it had a map table name to table.
Now it has nested map. Map of database name to map of table name to
table. This is because, different databases may have tables with the same
name. The same thing with MutableTablesState.

This change required to also support multiple databases in *TablesState classes.
No ImmutableTablesState and MutableTablesState are have additional
parameter in get method a database name. This is needed when you want to
access a table instance which with name which was used in several
databases.

Ticket: SWARM-884

Test Plan: running tempto examples

Reviewers: losipiuk, ksobczak

Pull request: https://github.com/prestodb/tempto/pull/25