Scan whole classpath in ReflectionHelper
Previously list of jars which are scanned for classes (mostly for sake
of auto binding) was limited to io.*. This limits users of tempto and
actualy prevented tempto from working in Presto after the prefix was
chaned from com.* to io.*.

It seems like scanning whole classpath, though not fast, completes in
reasonable amount of time. Around 10s on macbook pro.