Add GitHub Action for Gradle build in Java CI
Added a GitHub Action workflow to run Gradle build for the prestodb/tempto
repository. This action automates the CI process for Java projects and
ensures consistency in builds for every push and pull request to the master
branch.

PR: https://github.com/Dilli-Babu-Godari/tempto/pull/2
This PR serves as proof that the CI configuration functions correctly.