Upgrades for java 17 support
To be compatible with recent upgrades to airlift and presto for
Java 17, Tempto needs some upgrades in order to run against Presto
JVMs which have newer dependencies.

Guice needs to be upgraded to 7.0 which brings in utilization of
the jakarta namespaced injection and annotation dependencies.
Additionally, I neeeded to upgrade jsch and mina_sshd dependencies
in order to get the existing tests passing. This required some
refactoring in test classes